Added code to interpret input and execute instructions
Added the mars exploration class which takes in the sample input and works through it to create a mars instance and a robot instance for every robot coordinates and direction supplied. Added a validation and converter class to valid the input and return it in a usable way for Mars.cs and Robot.cs.